+++ wxc_sdk/telephony/devices/__init__.py
@@ -17,7 +17,7 @@ from ...api_child import ApiChild
 from ...base import ApiModel, plus1, to_camel, enum_str
 from ...base import SafeEnum as Enum
 from ...common import PrimaryOrShared, UserType, ValidationStatus, DeviceCustomization, IdAndName, \
-    ApplyLineKeyTemplateAction, UserLicenseType, DeviceType
+    ApplyLineKeyTemplateAction, UserLicenseType, DeviceType, UsageType
 
 __all__ = ['DeviceManufacturer', 'DeviceManagedBy', 'OnboardingMethod', 'DeviceSettingsConfiguration',
            'SupportsLogCollection', 'SupportedDevice', 'SupportedDevices', 'MemberCommon', 'DeviceMember',
@@ -697,6 +697,8 @@ class TelephonyDevicesApi(ApiChild, base='telephony/config'):
 
     def available_members(self, device_id: str, location_id: str = None, member_name: str = None,
                           phone_number: str = None, extension: str = None, org_id: str = None,
+                          usage_type: UsageType = None,
+                          order: str = None,
                           **params) -> Generator[AvailableMember, None, None]:
         """
         Search members that can be assigned to the device.
@@ -716,6 +718,11 @@ class TelephonyDevicesApi(ApiChild, base='telephony/config'):
         :type phone_number: str
         :param extension: Search (Contains) based on extension.
         :type extension: str
+        :param usage_type: Search for members eligible to become the owner of the device, or share line on the device.
+        :type usage_type: UsageType
+        :param order: Sort the list of available members on the device in ascending order by name, use either last name
+            `lname` or first name `fname`. Default: last name in ascending order.
+        :type order: str
         :param org_id: Retrieves the list of available members on the device in this Organization.
         :type org_id: str
         :return: list of available members